undo bkwards compat; appears working but needs addl testing
